vulkan::
 - better structures for buffer, memory

+++ gearoenix/vulkan/command/gx-vk-cmd-buffer.cpp
@@ -79,7 +79,7 @@ void gearoenix::vulkan::command::Buffer::copy(
     buffer::Buffer& des,
     const std::vector<VkBufferCopy>& info) noexcept
 {
-    vkCmdCopyBuffer(vulkan_data, src.get_vulkan_data(), des.get_vulkan_data(), info.size(), info.data());
+    vkCmdCopyBuffer(vulkan_data, src.get_vulkan_data(), des.get_vulkan_data(), static_cast<std::uint32_t>(info.size()), info.data());
 }
 
 void gearoenix::vulkan::command::Buffer::begin(const RenderPass& render_pass, const Framebuffer& framebuffer) noexcept
